Problem 9
The lack of expressiveness in C++ is pretty apparent in this one.  I definitely could make the performance of Haskell and Python better in this one by copying the mathematical optimisation in the C++ solution however I'm only really going for simplicity in those languages and I feel they're pretty simple to follow.